+++ src/runtime.rs
@@ -11,7 +11,7 @@ pub fn run_file(path: PathBuf, env: &mut Environment) -> bool {
     match std::fs::read_to_string(path) {
         Ok(prelude) => parse_and_eval(&prelude, env),
         Err(e) => {
-            eprintln!("\x1b[31m[ERROR]\x1b[0mFailed to read file:\n  {}", e);
+            eprintln!("\x1b[31m[IO ERROR]\x1b[0mFailed to read file:\n  {}", e);
             let _ = io::stderr().flush();
             false
         }
@@ -81,7 +81,7 @@ pub fn parse_and_eval(text: &str, env: &mut Environment) -> bool {
                 },
                 Err(e) => {
                     let _ = io::stdout().flush();
-                    eprintln!("\x1b[31m[ERROR]\x1b[0m {}", e);
+                    eprintln!("\x1b[31m[RUNTIME ERROR]\x1b[0m\n{}", e);
                     let _ = io::stderr().flush();
                 }
             }
@@ -120,7 +120,7 @@ pub fn parse_and_eval(text: &str, env: &mut Environment) -> bool {
         }
 
         Err(e) => {
-            eprintln!("[PARSE FAILED]\n{}", e);
+            eprintln!("\x1b[31m[PARSE ERROR]\x1b[0m\n{}", e);
             let _ = io::stderr().flush();
             // if line.is_empty() {
             //     eprintln!("{}", e);